# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LUXORA Shopping Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all products
 */
export const getProductsQueryLimitDefault = 50;

export const GetProductsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(getProductsQueryLimitDefault),
});

export const GetProductsResponseItem = zod.object({
  productId: zod.string(),
  title: zod.string(),
  price: zod.number(),
  description: zod.string(),
  category: zod.string(),
  images: zod.array(zod.string()),
  video: zod.string().nullish(),
  affiliateLink: zod.string(),
  clicks: zod.number().optional(),
  rating: zod.number().nullish(),
  totalReviews: zod.number().nullish(),
  createdAt: zod.string().nullish(),
});
export const GetProductsResponse = zod.array(GetProductsResponseItem);

/**
 * @summary Get trending products sorted by clicks
 */
export const getTrendingProductsQueryLimitDefault = 10;

export const GetTrendingProductsQueryParams = zod.object({
  limit: zod.coerce.number().default(getTrendingProductsQueryLimitDefault),
});

export const GetTrendingProductsResponseItem = zod.object({
  productId: zod.string(),
  title: zod.string(),
  price: zod.number(),
  description: zod.string(),
  category: zod.string(),
  images: zod.array(zod.string()),
  video: zod.string().nullish(),
  affiliateLink: zod.string(),
  clicks: zod.number().optional(),
  rating: zod.number().nullish(),
  totalReviews: zod.number().nullish(),
  createdAt: zod.string().nullish(),
});
export const GetTrendingProductsResponse = zod.array(
  GetTrendingProductsResponseItem,
);

/**
 * @summary Get featured products (newest + highest rated)
 */
export const GetFeaturedProductsResponseItem = zod.object({
  productId: zod.string(),
  title: zod.string(),
  price: zod.number(),
  description: zod.string(),
  category: zod.string(),
  images: zod.array(zod.string()),
  video: zod.string().nullish(),
  affiliateLink: zod.string(),
  clicks: zod.number().optional(),
  rating: zod.number().nullish(),
  totalReviews: zod.number().nullish(),
  createdAt: zod.string().nullish(),
});
export const GetFeaturedProductsResponse = zod.array(
  GetFeaturedProductsResponseItem,
);

/**
 * @summary Get a single product by ID
 */
export const GetProductParams = zod.object({
  productId: zod.coerce.string(),
});

export const GetProductResponse = zod.object({
  productId: zod.string(),
  title: zod.string(),
  price: zod.number(),
  description: zod.string(),
  category: zod.string(),
  images: zod.array(zod.string()),
  video: zod.string().nullish(),
  affiliateLink: zod.string(),
  clicks: zod.number().optional(),
  rating: zod.number().nullish(),
  totalReviews: zod.number().nullish(),
  createdAt: zod.string().nullish(),
});

/**
 * @summary Get reviews for a product
 */
export const GetProductReviewsParams = zod.object({
  productId: zod.coerce.string(),
});

export const GetProductReviewsResponseItem = zod.object({
  userId: zod.number(),
  productId: zod.string(),
  rating: zod.number(),
  createdAt: zod.string(),
});
export const GetProductReviewsResponse = zod.array(
  GetProductReviewsResponseItem,
);

/**
 * @summary Submit a product review
 */
export const SubmitReviewParams = zod.object({
  productId: zod.coerce.string(),
});

export const SubmitReviewBody = zod.object({
  userId: zod.number(),
  rating: zod.number(),
});

export const SubmitReviewResponse = zod.object({
  userId: zod.number(),
  productId: zod.string(),
  rating: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Get all product categories with product counts
 */
export const GetCategoriesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  productCount: zod.number(),
  icon: zod.string(),
});
export const GetCategoriesResponse = zod.array(GetCategoriesResponseItem);

/**
 * @summary Get all digital products (Luxora Learn)
 */
export const GetDigitalProductsResponseItem = zod.object({
  productId: zod.string(),
  title: zod.string(),
  price: zod.number(),
  description: zod.string(),
  fileLink: zod.string().nullish(),
  createdAt: zod.string().nullish(),
});
export const GetDigitalProductsResponse = zod.array(
  GetDigitalProductsResponseItem,
);

/**
 * @summary Create an order with delivery address
 */
export const createOrderBodyAddressCountryDefault = `India`;

export const CreateOrderBody = zod.object({
  items: zod.array(
    zod.object({
      productId: zod.string(),
      title: zod.string(),
      price: zod.number(),
      affiliateLink: zod.string(),
    }),
  ),
  address: zod.object({
    line1: zod.string(),
    line2: zod.string().nullish(),
    city: zod.string(),
    state: zod.string(),
    pincode: zod.string(),
    country: zod.string().default(createOrderBodyAddressCountryDefault),
  }),
  customerName: zod.string(),
  customerPhone: zod.string(),
  customerEmail: zod.string().nullish(),
  totalAmount: zod.number(),
});

/**
 * @summary Get platform stats (product count, categories, reviews)
 */
export const GetStatsResponse = zod.object({
  totalProducts: zod.number(),
  totalCategories: zod.number(),
  totalReviews: zod.number(),
  totalOrders: zod.number(),
});